Split workflow.ts: extract graph/affected and spec/expect commands
`src/cli/workflow.ts` had grown to 572 lines mixing 10 commands with
shared helpers. Following the existing `workflow/verify.ts` pattern:

- workflow/paths.ts: shared baseline / snapshot / report paths so the
  three submodules don't have to recompute VRT_ROOT.
- workflow/graph.ts: `runGraph` + `runAffected` (dep-graph commands).
- workflow/spec.ts: `runIntrospect` + `runSpecVerify` + `runExpect`
  (UI-spec authoring & expectation generation).
- workflow.ts: dispatcher (~320 lines, was 572). Init/capture/verify/
  approve/report stay inline since they all manage screenshot dirs.
